First iteration of hacker events
**Added hacker events to Blade.**
Hacker events are classified as normal events, except they have a hackathon id corresponding to their assigned hackathon.
While normal events are assigned to members, hacker events are assigned to the hacker attendees of each individual hackathon.
They are created using the same interface for normal events, except a dropdown to select the attached hackathon was added, defaulting to None.
No matter the settings, if you attach a hackathon to an event, that event WILL be hacker only.
This also means that hackathon events can use ANY event tag and still work, as long as they have that attached id.
This will hopefully make data analysis after the hackathon easier.

The Club/Events screen has been updated to match the existence of hacker events.
Buttons for each usertype check in have been added, which open the existing QR scanner but with different event selections.
Member codes cannot be used for hacker events, and hacker codes cannot be used for normal events.
To resolve issues during the event, the user's QR code will now outwardly display if it is a hacker or a member code.

The Club/Data screen includes hackathon events since they are part of the same DB table, so I also added a checkbox to include/exclude these events from the statistics.
I also added an Event data section to the Hackathon/Data screen to look at event statisitcs for each hackathon specifically.
Right now, it's identical to the original event data screen besides the selector dropdown, we can change this later if need be.

*Note: check-in still uses the old system and NOT the hackathon event system, this will be added later.*